requirements.txt
agentforce-sdk
salesforce-cli
pyyaml
